fixed hib user

+++ src/main/java/ua/com/startup/investments/entities/User.java
@@ -63,7 +63,7 @@ public class User {
      * @param phone display phone number of user
      * @param role display role of user
      */
-    public User(long id, String name, String surname, String address, Integer phone, Roles role) {
+    public User(int id, String name, String surname, String address, Integer phone, Roles role) {
         this.id = id;
         this.name = name;
         this.surname = surname;
@@ -72,14 +72,14 @@ public class User {
         this.role = role;
     }
 
-    public User(long id, String name, String surname, Integer phone) {
+    public User(int id, String name, String surname, Integer phone) {
         this.id = id;
         this.name = name;
         this.surname = surname;
         this.phone = phone;
     }
 
-    public User(long id, String name) {
+    public User(int id, String name) {
         this.id = id;
         this.name = name;
     }
@@ -103,7 +103,7 @@ public class User {
     /**
      * Getters and setters methods by all fields of user.
      */
-    public long getId() {
+    public int getId() {
         return id;
     }
